# co-shop-core/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Group Groceries API
  version: 1.0.1
  description: 'Mobile-first cooperative shopping API for group trips (e.g., ski huts).

    Covers groups, members, products, items (products in a group), proposals (quantity
    suggestions),

    commits (who brings what), and requests (member wants how much). Includes compute
    endpoints for

    planning quantities and fair settlement.

    '
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local Dev
...
security:
  - bearerAuth: [ ]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GroupIdParam:
      in: path
      name: groupId
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Group:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        inviteCode:
          type: string
    GroupCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        members:
          type: array
          items:
            type: integer
            format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    Member:
      type: object
      required:
        - id
        - displayName
      properties:
        id:
          type: integer
          format: int64
        displayName:
          type: string
...
paths:
  /groups:
    post:
      tags:
        - Groups
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
      operationId: createGroup
  /groups/{groupId}:
    get:
      tags:
        - Groups
      summary: Get group
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      responses:
        '200':
          description: Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Not found
      operationId: getGroup
  /groups/{groupId}/members:
    get:
      tags:
        - groups
      summary: List members of a group
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
      operationId: listGroupMembers
  /groups/{groupId}/settlement:
    get:
      tags:
        - Groups
      summary: Compute fair settlement for a group
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      responses:
        '200':
          description: Settlement
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      operationId: computeGroupSettlement
  /groups/join:
    post:
      tags:
        - Groups
      summary: Join a group via invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inviteCode
                - memberId
              properties:
                inviteCode:
                  type: string
                memberId:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Joined
        '400':
          description: Invalid invite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: joinGroup
...